trading_ai/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
yfinance>=0.2.30
ccxt>=4.0.0
ta>=0.10.2
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0
flask>=3.0.0
plotly>=5.18.0
matplotlib>=3.8.0
